+++ battlecode22-scaffold/src/ATopTenBot/BotArchon.java
@@ -59,6 +59,15 @@ public class BotArchon extends Util{
         atTargetLocationForTransform = false;
         transformAndMoveOrigin = null;
         goodPlace = false;
+        // if(BIRTH_ROUND % 3 == 0) {
+            // currentDestination = ratioPointBetweenTwoMapLocations(parentArchonLocation, rememberedEnemyArchonLocations[0], 0.15);
+        // } 
+        // else if (BIRTH_ROUND % 3 == 1){
+        currentDestination = ratioPointBetweenTwoMapLocations(parentArchonLocation, rememberedEnemyArchonLocations[1], 0.15);
+        // }
+        // else{
+            // currentDestination = ratioPointBetweenTwoMapLocations(parentArchonLocation, rememberedEnemyArchonLocations[2], 0.15);
+        // }
     }